cborrs::cbordet

Type Alias CborDet

pub type CborDet<'a> = cbordet<'a>;

An abstract type to represent a Deterministic CBOR object, which can be obtained either using “constructor” functions of this module, or by parsing from serialized data. Declared in pulse/CBOR.Pulse.Det.API.Rust.fsti. Its corresponding pure (“mathematical”) functional specification is the cbor type in spec/CBOR.Spec.Type.fsti.

Aliased Type

pub enum CborDet<'a> {}